Move CanvasContextSpy to a place where it can be exported.
