+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './core';
 export * from './connected';
+export * from './test-util';
