+++ src/test-util/index.ts
@@ -1,0 +1,1 @@
+export { default as CanvasContextSpy, PropertySet, MethodCall } from './CanvasContextSpy';
